# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking professional achievements and trophies.",
        "title": "Professional Portfolio Tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/achievements": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Get list of achievements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Achievement"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "achievements"
                ],
                "summary": "Create new achievement",
                "parameters": [
                    {
                        "description": "Achievement to create",
                        "name": "achievement",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Achievement"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Achievement"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/trophies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trophies"
                ],
                "summary": "Get list of trophies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Trophy"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trophies"
                ],
                "summary": "Create new trophy",
                "parameters": [
                    {
                        "description": "Trophy to create",
                        "name": "trophy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Trophy"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Trophy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Achievement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "impact": {
                    "type": "string"
                },
                "skill_used": {
                    "type": "string"
                }
            }
        },
        "main.Trophy": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}